Add ability to loop one-shot animations.
At the same time, remove Xebow.Animation.{Loop, OneShot}, these
animation types can now both be handled by Xebow.Animation.Static.
